requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pyautogui>=0.9.54
pygetwindow>=0.0.9
pytesseract>=0.3.10
Pillow>=10.0.0
psutil>=5.9.0
pyperclip>=1.8.2 